restructure recommendation page and related components
