+++ src/App.jsx
@@ -5,7 +5,7 @@ import 'normalize.css'
 import Typography from './styles/Typography'
 import GlobalStyles from './styles/GlobalStyles'
 import Layout from './components/layout/layout-wrapper'
-import {ErrorMessage, FullPageSpinner} from './components/lib'
+
 
 import authRouter from './utils/login'
 import storage from './utils/storage'
@@ -13,6 +13,8 @@ import {ErrorBoundary} from 'react-error-boundary'
 import AuthenticatedApp from './authenticated-app'
 import UnauthenticatedApp from './unauthenticated-app'
 import {useAsync} from './utils/hooks'
+import { FullPageSpinner, ErrorMessage } from 'components/lib'
+
 
 async function getUser() {
   let user = null
