+++ src/authenticated-app.jsx
@@ -1,10 +1,9 @@
 import { Switch, Route } from 'react-router-dom'
+import PropTypes from 'prop-types'
 import Home from 'pages/home'
-import AddRecommendation from 'pages/AddRecommendation'
-import Recommendations from 'pages/Reccomendations'
+import Recommendations from 'pages/recommendations'
 import Watchlist from 'pages/watchlist'
 import Search from 'pages/search/search'
-import PropTypes from 'prop-types'
 import Recommendation from 'pages/recommendation/recommendation'
 
 function AuthenticatedApp({ user }) {
@@ -18,9 +17,6 @@ function AuthenticatedApp({ user }) {
       </Route>
       <Route path='/recommendations'>
         <Recommendations />
-      </Route>
-      <Route path='/add-recommendation'>
-        <AddRecommendation user={user} />
       </Route>
       <Route path='/recommendation/:id' >
         <Recommendation user={user} />
