--- src/pages/AddRecommendation.js
@@ -1,15 +1,0 @@
-import AddRecommendationTable from '../components/AddRecommendationTable'
-import { SectionStyles } from '../styles/styles'
-
-const AddRecommendation = () => {
-
-  return(
-    <SectionStyles>
-      <section>
-        <AddRecommendationTable />
-      </section>
-    </SectionStyles>
-  )
-}
-
-export default AddRecommendation
