+++ src/pages/search/components/search-modal.jsx
@@ -3,8 +3,9 @@ import PropTypes from 'prop-types'
 import { useAddRecommendation } from 'utils/recommendations'
 import { useIndividualSearch } from 'utils/search'
 import { SearchResultSkeleton } from 'utils/skeleton'
-import { ErrorMessage } from 'components/lib'
+
 import Modal from 'components/modals/modal-wrapper'
+import { ErrorMessage } from 'components/lib'
 
 const SearchModalStyles = styled.div`
   box-sizing: border-box;
